+++ tests/testthat/test-mcmc-resid.R
@@ -28,7 +28,7 @@ test_that("Test that MCMC residuals are working", {
   resid_mle <- residuals(m1)
   resid_mcmc <- residuals(m1, type = "mle-mcmc", mcmc_iter = 1000)
   expect_lt(abs(mean(resid_mcmc)), 0.2)
-  expect_equal(sd(resid_mcmc), 1.000, tolerance = 1e-2)
+  expect_equal(sd(resid_mcmc), 1, tolerance = 0.02)
   expect_equal(cor(resid_mcmc, resid_mle), 0.955, tolerance = 1e-2)
 
   # binomial example from scratch/stan-testing